refactor : add visible interface and implemented classes, and use it by composite pattern
